chore: refactor to reduce duplicated code
