+++ cmake/libraries/crypto.cmake
@@ -1,8 +1,6 @@
 file(GLOB_RECURSE CRYPTO_SOURCES
     ./include/crypto/*.h
-    ./include/crypto/encoding/*.h
     ./src/crypto/*.c
-    ./src/crypto/encoding/*.c
 )
 
 add_library(libcrypto
@@ -11,6 +9,7 @@ add_library(libcrypto
 )
 
 target_compile_options(libcrypto PRIVATE ${flags})
+target_link_libraries(libcrypto libencoding)
 target_link_libraries(libcrypto libprotobuf)
 target_link_libraries(libcrypto libmultihash)
 target_link_libraries(libcrypto mbedtls)
