+++ cmake/libraries/multiaddr.cmake
@@ -1,13 +1,11 @@
 add_library(libmultiaddr
     SHARED
-        ./include/multiaddr/base58.h
         ./include/multiaddr/endian.h
         ./include/multiaddr/multiaddr.h
         ./include/multiaddr/protocols.h
         ./include/multiaddr/protoutils.h
         ./include/multiaddr/varhexutils.h
         ./include/multiaddr/varint.h
-        ./src/multiaddr/base58.c
         ./src/multiaddr/multiaddr.c
         ./src/multiaddr/protocols.c
         ./src/multiaddr/protoutils.c
@@ -16,6 +14,7 @@ add_library(libmultiaddr
 )
 target_compile_options(libmultiaddr PRIVATE ${flags})
 target_link_libraries(libmultiaddr m)
+target_link_libraries(libmultiaddr libencoding)
 
 add_executable(libmultiaddr-test ./src/multiaddr/multiaddr_test.c)
 target_link_libraries(libmultiaddr-test cmocka)
