+++ cmake/libraries/multibase.cmake
@@ -3,6 +3,6 @@ add_library(libmultibase
         ./include/multibase/multibase.h
         ./src/multibase/multibase.c
 )
-target_compile_options(libmultibase PRIVATE ${flags})
 
-target_link_libraries(libmultibase libcrypto)
+target_compile_options(libmultibase PRIVATE ${flags})
+target_link_libraries(libmultibase libencoding)
