+++ include/encoding/base16.h
