+++ include/encoding/base32.h
